Update index.js
Added error handling: We added a middleware function that catches any errors that occur during the request-response cycle and sends a response with a status code of 500 and the message "Something went wrong!". This will help to provide a better user experience by letting users know that something has gone wrong, rather than leaving them to wonder why they aren't getting a response.

Used a configuration management library: We replaced the use of environment variables with the config library to manage configuration values. This will make it easier to manage and change configuration values, as they will be stored in a central location and can be accessed using the config library.

Used a process manager: We used the pm2 library to start the server and ensure that it is always running. If the server stops for any reason, pm2 will automatically restart it. This will help to ensure that the server is always available to handle requests.

Added logging: We added logging using the winston library. This will allow us to log important events and messages, which can be useful for debugging issues or understanding what's happening in the application.