Update kamyroll.js
Replaced the axios import with a require statement: In the original code, axios was imported using the import syntax, which is not supported in Node.js. To fix this, I replaced the import statement with a require statement, which is the correct way to import modules in Node.js.

Removed the updateKitsuMapping function: In the original code, the updateKitsuMapping function was used to fetch and store a mapping of IMDb IDs to Kitsu IDs from a GitHub repository. However, this function is not being called anywhere in the code, so it is unnecessary.

Removed the hasKitsuMapping function: In the original code, the hasKitsuMapping function was used to check if a given IMDb ID had a corresponding Kitsu ID in the kitsuMapping object. However, this object is not being used anywhere in the code, so this function is unnecessary.

Removed the getTitles function: In the original code, the getTitles function was used to fetch the titles for a given Kitsu ID from the Kitsu API. However, this information is not being used anywhere in the code, so this function is unnecessary.

Removed the getSeasonCrunchyrollId function: In the original code, the getSeasonCrunchyrollId function was used to fetch the Crunchyroll ID for a given Kitsu ID from the Kitsu API. However, this information is not being used anywhere in the code, so this function is unnecessary.

Imported the memory-cache library: To implement caching in the kamyroll.js module, I imported the memory-cache library at the top of the file using the require statement.

Added a cache key variable: To store the results of the getKitsuId function in the cache, I added a variable called cacheKey that concatenates the `im

Checked if the result is already in the cache: Before making the HTTP request to the Kitsu API, I checked if the result is already in the cache using the cache.get method. If the result is found in the cache, I returned it immediately.

Cached the result: After the HTTP request to the Kitsu API has completed, I stored the result in the cache using the cache.put method. I specified the cacheKey variable as the key, the result as the value, and a timeout of 1 hour (3600 seconds) as the third argument. This will cause the result to be stored in the cache for 1 hour before it is automatically removed.